test: wrapper around GPModel to reach simple and clean interface
